Add form deleting event

+++ src/Forms/Form.php
@@ -12,6 +12,7 @@ use Statamic\Events\FormBlueprintFound;
 use Statamic\Events\FormCreated;
 use Statamic\Events\FormCreating;
 use Statamic\Events\FormDeleted;
+use Statamic\Events\FormDeleting;
 use Statamic\Events\FormSaved;
 use Statamic\Events\FormSaving;
 use Statamic\Facades\Blueprint;
@@ -220,6 +221,10 @@ class Form implements Arrayable, Augmentable, FormContract
      */
     public function delete()
     {
+        if (FormDeleting::dispatch($this) === false) {
+            return false;
+        }
+
         $this->submissions()->each->delete();
 
         File::delete($this->path());
